# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/vehicles": {
      "get": {
        "summary": "Get all vehicles with filters",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicles retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Add vehicle to inventory",
        "tags": [
          "Vehicles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Vehicle added successfully"
          }
        }
      }
    }
  },
...
}